Added a ChartDisplayer Class
Chart Displayer Class will allow you to view and interact with the colour chart. clicking on one of the colours in the chart will show you the key frame that the colour is from.

to be used after processing. so you no longer have to process an image to view the colour chart and its respective frames.

Note: This project is messy. it is just intended to entertain me, i am too busy to ensure the code is written well in any way shape  or form. i only care about the output for now.